Moved SendInvoiceParams into fizz-kidz
Used the `callFirebaseFunction` method for sending an invoice.
Also cleaned up folder structure for utilities.
