+++ firebase/src/components/Forms/ExistingBookingForm/index.js
@@ -19,7 +19,7 @@ import Fab from '@material-ui/core/Fab'
 import { green, red } from '@material-ui/core/colors'
 import { validateFormOnChange, validateFormOnSubmit, errorFound } from '../validation'
 import { Booking } from 'fizz-kidz'
-import { capitalise } from '../../../utilities'
+import { capitalise } from '../../../utilities/stringUtilities'
 import { compose } from 'recompose'
 import withErrorDialog from '../../Dialogs/ErrorDialog'
 import WithConfirmationDialog from '../../Dialogs/ConfirmationDialog'
