+++ firebase/src/utilities/stringUtilities.ts
